wip: refactor time capsule item ui
